refine code, rename files, minor fixes
