+++ src/controllers/auth/login.js
@@ -1,0 +1,21 @@
+import { ONE_DAY } from '../../constants/index.js';
+import { loginUser } from '../../services/auth/login.js';
+
+export const loginUserController = async (req, res) => {
+  const session = await loginUser(req.body);
+  res.cookie('refreshToken', session.refreshToken, {
+    httpOnly: true,
+    expires: new Date(Date.now() + ONE_DAY),
+  });
+  res.cookie('sessionId', session._id, {
+    httpOnly: true,
+    expires: new Date(Date.now() + ONE_DAY),
+  });
+  res.json({
+    status: 200,
+    message: 'Successfully logged in an user!',
+    data: {
+      accessToken: session.accessToken,
+    },
+  });
+};
