+++ src/controllers/auth/register.js
@@ -1,4 +1,4 @@
-import { registerUser } from '../../services/auth/user.auth.js';
+import { registerUser } from '../../services/auth/user.js';
 
 export const registerUserController = async (req, res) => {
   const user = await registerUser(req.body);
