+++ src/routers/index.js
@@ -3,11 +3,9 @@
 import { Router } from 'express';
 import authRouter from './auth.js';
 import usersRouter from './users.js';
-//import routers
 
 const router = Router();
 
-//add here router.use();
 router.use('/api/auth', authRouter);
 router.use('/api/users', usersRouter);
