+++ src/server.js
@@ -8,6 +8,7 @@ import { getEnvVar } from './utils/getEnvVar.js';
 import { notFoundHandler } from './middlewares/notFoundHandler.js';
 import { errorHandler } from './middlewares/errorHandler.js';
 import cookieParser from 'cookie-parser';
+
 const PORT = Number(getEnvVar('PORT', '3000'));
 
 export const startServer = () => {
