--- src/services/auth/login.auth.js
@@ -1,27 +1,0 @@
-import createHttpError from 'http-errors';
-import User from "../../db/models/auth/user.auth.js";
-import { FIFTEEN_MINUTES, ONE_DAY } from '../../constants/index.js';
-import Session  from '../../db/models/auth/session.auth.js';
-import bcrypt from 'bcrypt';
-import { randomBytes } from 'crypto';
-export const loginUser = async (payload) => {
-    const user = await User.findOne({ email: payload.email });
-    if (!user) {
-        throw createHttpError(401, "User not found");
-    }
-    const isEqual = await bcrypt.compare(payload.password, user.password);
-    if (!isEqual) {
-        throw createHttpError(401, 'Unauthorized');
-    }
-    await Session.deleteOne({ userId: user._id });
-    const accessToken = randomBytes(30).toString('base64');
-    const refreshToken = randomBytes(30).toString('base64');
-
-    return await Session.create({
-        userId: user._id,
-        accessToken,
-        refreshToken,
-        accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
-        refreshTokenValidUntil: new Date(Date.now() + ONE_DAY),
-    });
-};
