+++ src/services/auth/user.js
@@ -1,6 +1,6 @@
 import bcrypt from 'bcrypt';
 import createHttpError from 'http-errors';
-import User from '../../db/models/auth/user.auth.js';
+import User from '../../db/models/auth/user.js';
 
 export const registerUser = async (payload) => {
   const existingUser = await User.findOne({ email: payload.email });
